+++ app/src/main/java/com/example/playlistmaker/ActivitySettings.kt
@@ -1,20 +1,14 @@
-package com.example.playlistmaker.activity
+package com.example.playlistmaker
 
-import android.annotation.SuppressLint
 import android.content.ActivityNotFoundException
 import android.content.Intent
-import android.content.SharedPreferences
+import android.net.Uri
 import android.os.Bundle
 import android.widget.ImageButton
 import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.SwitchCompat
-import androidx.core.net.toUri
-import com.example.playlistmaker.App
-import com.example.playlistmaker.MY_SAVES
-import com.example.playlistmaker.R
-import com.example.playlistmaker.THEME
 
 class ActivitySettings : AppCompatActivity() {
 
@@ -23,16 +17,12 @@ class ActivitySettings : AppCompatActivity() {
     private lateinit var shareApp:TextView
     private lateinit var support:TextView
     private lateinit var userAgreement:TextView
-    private lateinit var sharedPreferences: SharedPreferences
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_settings)
-        sharedPreferences = getSharedPreferences(MY_SAVES,MODE_PRIVATE)
         initViews()
-        switch.isChecked = sharedPreferences.getBoolean(THEME,false)
         setOnClick()
-
     }
 
     private fun initViews() {
@@ -42,7 +32,6 @@ class ActivitySettings : AppCompatActivity() {
         support = findViewById(R.id.activity_settings_textView_write_to_support)
         userAgreement = findViewById(R.id.activity_settings_textView_user_agreement)
     }
-    @SuppressLint("QueryPermissionsNeeded")
     private fun setOnClick(){
         support.setOnClickListener {
             val email = getString(R.string.email)
@@ -50,7 +39,7 @@ class ActivitySettings : AppCompatActivity() {
             val text = getString(R.string.text_mail)
             val intent = Intent().apply {
                 action = Intent.ACTION_SENDTO
-                data = "mailto:".toUri()
+                data = Uri.parse("mailto:")
                 putExtra(Intent.EXTRA_EMAIL, arrayOf(email))
                 putExtra(Intent.EXTRA_SUBJECT, subject)
                 putExtra(Intent.EXTRA_TEXT, text)
@@ -81,16 +70,18 @@ class ActivitySettings : AppCompatActivity() {
         userAgreement.setOnClickListener {
             val intent = Intent().apply {
                 action = Intent.ACTION_VIEW
-                data = getString(R.string.link_practicum_offer).toUri()
+                data = Uri.parse(getString(R.string.link_practicum_offer))
             }
             startActivity(intent)
         }
-        buttonBack.setOnClickListener {
-            finish()
-        }
-        switch.setOnCheckedChangeListener { switcher, checked ->
-            (applicationContext as App).switchTheme(checked)
-        }
 
+        buttonBack.setOnClickListener {
+            startActivity(Intent(this, MainActivity::class.java))
+        }
+        switch.setOnClickListener { // TODO:   }
+            support.setOnClickListener {
+                // TODO:
+            }
+        }
     }
 }
