+++ app/src/main/java/com/example/playlistmaker/MainActivity.kt
@@ -1,10 +1,9 @@
-package com.example.playlistmaker.activity
+package com.example.playlistmaker
 
 import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
 
 class MainActivity : AppCompatActivity() {
     private lateinit var buttonSearch: Button
