+++ app/src/main/java/com/example/playlistmaker/Track.kt
@@ -1,6 +1,6 @@
-package com.example.playlistmaker.data_classes
+package com.example.playlistmaker
 
 data class Track(val trackName: String,
                  val artistName:String,
-                 val trackTimeMillis: Long,
+                 val trackTime: String,
                  val artworkUrl100:String)
