--- app/src/main/java/com/example/playlistmaker/constants.kt
@@ -1,12 +1,0 @@
-package com.example.playlistmaker
-
-const val MY_SAVES = "my saves"
-const val THEME = "theme"
-const val LIST_TRACKS = "list tracks"
-const val ENTERED_TEXT = "ENTERED_TEXT"
-const val RECYCLER_VISIBILITY = "RECYCLER_VISIBILITY"
-const val IMAGE_ERROR_VISIBILITY = "IMAGE_ERROR_VISIBILITY"
-const val IMAGE_ERROR_IMAGE_RESOURCE = "IMAGE_ERROR_IMAGE_RESOURCE"
-const val TEXT_ERROR_VISIBILITY = "TEXT_ERROR_VISIBILITY"
-const val TEXT_ERROR_RESOURCE = "TEXT_ERROR_RESOURCE"
-const val BUTTON_UPDATE_VISIBILITY = "BUTTON_UPDATE_VISIBILITY"
