--- app/src/main/java/com/example/playlistmaker/data_classes/TrackResponse.kt
@@ -1,5 +1,0 @@
-package com.example.playlistmaker.data_classes
-
-data class TrackResponse(
-    val results: List<Track>
-)
