--- app/src/main/res/drawable/cursor_drawable.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <size android:width="2dp"/>
-    <solid android:color="@color/blue"/>
-</shape>
