--- app/src/main/res/drawable/no_tracks.xml
@@ -1,22 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="120dp"
-    android:height="120dp"
-    android:viewportWidth="120"
-    android:viewportHeight="120">
-  <path
-      android:pathData="M34,29.5L111,3V31.5L40,42.1V76.5V79H39.89C38.95,89.14 32.42,97 24.5,97C15.94,97 9,87.82 9,76.5C9,65.18 15.94,56 24.5,56C28.08,56 31.38,57.6 34,60.3V43V35V29.5Z"
-      android:fillColor="@color/iconColorBlackWhite"
-      android:fillType="evenOdd"/>
-  <path
-      android:pathData="M71.5,62.5m-34.5,0a34.5,34.5 0,1 1,69 0a34.5,34.5 0,1 1,-69 0"
-      android:fillColor="#1A1B22"/>
-  <path
-      android:pathData="M71.5,28C62.35,28 53.57,31.63 47.1,38.1C40.63,44.57 37,53.35 37,62.5C37,71.65 40.63,80.43 47.1,86.9C53.57,93.37 62.35,97 71.5,97C80.65,97 89.43,93.37 95.9,86.9C102.36,80.43 106,71.65 106,62.5C106,53.35 102.36,44.57 95.9,38.1C89.43,31.63 80.65,28 71.5,28ZM65.03,56.03C65.03,57.17 64.58,58.27 63.77,59.08C62.96,59.89 61.86,60.34 60.72,60.34C59.58,60.34 58.48,59.89 57.67,59.08C56.86,58.27 56.41,57.17 56.41,56.03C56.41,54.89 56.86,53.79 57.67,52.98C58.48,52.17 59.58,51.72 60.72,51.72C61.86,51.72 62.96,52.17 63.77,52.98C64.58,53.79 65.03,54.89 65.03,56.03ZM82.28,51.72C83.43,51.72 84.52,52.17 85.33,52.98C86.14,53.79 86.59,54.89 86.59,56.03C86.59,57.17 86.14,58.27 85.33,59.08C84.52,59.89 83.43,60.34 82.28,60.34C81.14,60.34 80.04,59.89 79.23,59.08C78.42,58.27 77.97,57.17 77.97,56.03C77.97,54.89 78.42,53.79 79.23,52.98C80.04,52.17 81.14,51.72 82.28,51.72ZM83.59,71.13H87.03C87.6,71.13 88.15,71.35 88.55,71.76C88.95,72.16 89.18,72.71 89.18,73.28C89.18,73.85 88.95,74.4 88.55,74.81C88.15,75.21 87.6,75.44 87.03,75.44H83.59C78.94,75.44 74.4,76.81 70.54,79.39C70.3,79.55 70.04,79.67 69.75,79.73C69.47,79.8 69.18,79.8 68.9,79.75C68.61,79.7 68.34,79.59 68.1,79.43C67.86,79.27 67.66,79.06 67.5,78.82C67.34,78.58 67.23,78.31 67.17,78.03C67.12,77.75 67.12,77.46 67.18,77.17C67.24,76.89 67.36,76.62 67.53,76.39C67.69,76.15 67.9,75.95 68.15,75.8C72.72,72.75 78.09,71.13 83.59,71.13Z"
-      android:fillColor="#FCD452"/>
-  <path
-      android:pathData="M80,95.5a15.5,20.5 0,1 0,31 0a15.5,20.5 0,1 0,-31 0z"
-      android:fillColor="@color/iconColorBlackWhite"/>
-  <path
-      android:pathData="M105,18.21L111,6V97H105V18.21Z"
-      android:fillColor="@color/iconColorBlackWhite"/>
-</vector>
