+++ app/src/main/res/layout/activity_search.xml
@@ -4,7 +4,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.ActivitySearch"
+    tools:context=".ActivitySearch"
     android:orientation="vertical">
 
     <LinearLayout
@@ -48,15 +48,13 @@
             android:drawableStart="@drawable/magnifier"
             android:drawablePadding="4dp"
             android:hint="@string/search"
-            android:imeOptions="actionDone"
             android:inputType="text"
             android:maxLines="1"
             android:paddingHorizontal="14dp"
             android:paddingVertical="@dimen/_8dp"
-            android:singleLine="true"
             android:textColor="@color/black"
             android:textColorHint="@color/activitySearch.grayBlack"
-            android:textCursorDrawable="@drawable/cursor_drawable"
+            android:textCursorDrawable="@color/blue"
             tools:ignore="TouchTargetSizeCheck">
 
         </EditText>
@@ -77,18 +75,6 @@
 
     </FrameLayout>
 
-    <TextView
-        android:id="@+id/activity_search_TextView_You_were_looking"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/You_were_looking"
-        android:layout_marginTop="@dimen/_18dp"
-        android:layout_marginBottom="@dimen/_12dp"
-        android:fontFamily="@font/ys_display_medium"
-        android:textSize="18sp"
-        android:gravity="center"
-        android:visibility="gone">
-    </TextView>
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/activity_search_recyclerView"
         android:layout_marginTop="@dimen/_16dp"
@@ -96,48 +82,5 @@
         android:layout_height="wrap_content">
 
     </androidx.recyclerview.widget.RecyclerView>
-    <Button
-        android:id="@+id/activity_search_clear_history"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="@string/clear_history"
-        android:textSize="14sp"
-        android:fontFamily="@font/ys_display_regular"
-        android:textColor="@color/iconColorWhiteBlack"
-        android:backgroundTint="@color/iconColorBlackWhite"
-        android:visibility="gone">
 
-    </Button>
-    <ImageView
-        android:id="@+id/activity_search_imageView_Error"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:src = "@drawable/no_internet"
-        android:layout_marginTop="100dp"
-        android:layout_gravity="center"
-        android:visibility="gone">
-    </ImageView>
-    <TextView
-        android:id="@+id/activity_search_textView_Error"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/no_internet"
-        android:textSize="19sp"
-        android:layout_marginTop="10dp"
-        android:layout_gravity="center"
-        android:gravity="center"
-        android:visibility="gone">
-    </TextView>
-    <Button
-        android:id="@+id/activity_search_button_update"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="Обновить"
-        android:layout_marginTop="24dp"
-        android:backgroundTint="@color/iconColorBlackWhite"
-        android:visibility="gone">
-
-    </Button>
 </androidx.appcompat.widget.LinearLayoutCompat>
