+++ app/src/main/res/layout/activitylibrary.xml
@@ -5,6 +5,6 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.ActivityLibrary">
+    tools:context=".ActivityLibrary">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
