+++ app/src/main/res/layout/track_view.xml
@@ -8,8 +8,7 @@
         android:id="@+id/track_view_logo"
         style="@style/style_track_view_logo"
         android:background="@drawable/rounded_background_2dp"
-        android:contentDescription="@string/logo"
-        android:layout_marginVertical="@dimen/_8dp">
+        android:contentDescription="@string/logo">
 
     </ImageButton>
