+++ app/src/main/res/values-night/colors.xml
@@ -1,21 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <color name="iconColorBlackWhite">#FFFFFFFF</color>
-    <color name="iconColorWhiteBlack">#1A1B22</color>
+    <color name="iconColorWhiteBlack">#FF000000</color>
     <color name="iconColorGrayWhite">#FFFFFFFF</color>
-    <color name="activitySearch.grayBlack">#1A1B22</color>
+    <color name="activitySearch.grayBlack">#FF000000</color>
     <color name="iconColorGrayBlack">#1A1B22</color>
     <color name="activitySearch.backgroundGrayWhite">#FFFFFFFF</color>
     <color name="white">#FFFFFFFF</color>
-    <color name="black">#1A1B22</color>
+    <color name="black">#FF000000</color>
     <color name="gray">#AEAFB4</color>
     <color name="blue">#3772E7</color>
-
     <color name="switch_thumb_inactive_color">#AEAFB4</color>
     <color name="switch_thumb_active_color">#3772E7</color>
     <color name="switch_track_inactive_color">#E6E8EB</color>
     <color name="switch_track_active_color">#9FBBF3</color>
-
     <color name="purple_200">#FFBB86FC</color>
     <color name="purple_500">#FF6200EE</color>
     <color name="purple_700">#FF3700B3</color>
