+++ app/src/main/res/values/styles.xml
@@ -76,9 +76,4 @@
         <item name="android:textSize">12sp</item>
         <item name="android:textColor">@color/iconColorGrayWhite</item>
     </style>
-
-    <style name="SwitchView">
-        <item name="android:thumbTint">@color/switch_thumb_color</item>
-        <item name="android:trackTint">@color/switch_track_color</item>
-    </style>
 </resources>
